Criação de arquivo de cabeçalho e teste de inserção de função externa
